requirements.txt
# HTTP requests
requests>=2.28.0

# Google API
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0

# Date handling
python-dateutil>=2.8.0

# Configuration management
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0

# Code quality
black>=23.0.0
flake8>=6.0.0